requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
jinja2>=3.1.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.0
numpy>=1.21.0
tqdm>=4.62.0
torch>=1.9.0